move generate-text in next.js examples to new route
