+++ examples/example-instrumentation-nextjs-v4/src/app/generate-text/loading.tsx
@@ -1,0 +1,7 @@
+export default async function Loading() {
+  return (
+    <div>
+      <p>Loading...</p>
+    </div>
+  );
+}
